autorun: fixup 'su -l' handling
When 'su' invokes a login shell with the '-' parameter the shell
will be reading /etc/profile, causing errors as the rapido profile
really is a system initialisation. So add another file 'vm_shell.env'
to be copied into /etc/profile and have vm_autorun.env to be
invoked from the 'real' login shell only.

Signed-off-by: Hannes Reinecke <hare@suse.de>